moved TodosPage to pages folder per assignment requirement
